Fix bug with product_ids, update checkout logic
