+++ my-app/db.json
@@ -1,78 +1,16 @@
 {
   "coffee": [
     {
-      "id": "3",
-      "product_id": "1",
-      "product_type": "coffee",
-      "brand_name": "Lavazza",
-      "description": "Lavazza Super Crema Whole Bean Coffee Blend, 2.2lb",
-      "image": "https://m.media-amazon.com/images/I/51MMPznBM1L._AC_UL320_.jpg",
-      "price": 23.99,
-      "in_stock_qty": 10
-    },
-    {
-      "id": "7",
-      "product_id": "2",
-      "product_type": "coffee",
-      "brand_name": "Death Wish Coffee",
-      "description": "Death Wish Coffee Co. Whole Bean Coffee, 1lb",
-      "image": "https://m.media-amazon.com/images/I/61+OtZZtI4L._AC_UL320_.jpg",
-      "price": 19.99,
-      "in_stock_qty": 12
-    },
-    {
-      "id": "9",
-      "product_id": "3",
-      "product_type": "coffee",
-      "brand_name": "Blue Bottle",
-      "description": "Blue Bottle Coffee, Bella Donovan, Whole Bean, 12oz",
-      "image": "https://m.media-amazon.com/images/I/31vGBP7UAoL._AC_UL320_.jpg",
-      "price": 18.99,
-      "in_stock_qty": 6
-    },
-    {
-      "id": "13",
-      "product_id": "5",
-      "product_type": "coffee",
-      "brand_name": "Stumptown",
-      "description": "Stumptown Coffee Roasters, Hair Bender, Whole Bean, 12oz",
-      "image": "https://m.media-amazon.com/images/I/71Q129rCmuL._AC_UL320_.jpg",
-      "price": 16.99,
-      "in_stock_qty": 7
-    },
-    {
-      "id": "15",
-      "product_id": "6",
-      "product_type": "coffee",
-      "brand_name": "Intelligentsia",
-      "description": "Intelligentsia Coffee, Black Cat Classic Espresso, Whole Bean, 12oz",
-      "image": "https://m.media-amazon.com/images/I/61ljzJa9oAL._AC_UL320_.jpg",
-      "price": 16,
-      "in_stock_qty": 6
-    },
-    {
-      "id": "23",
-      "product_id": "9",
-      "product_type": "coffee",
-      "brand_name": "Seattle's Best",
-      "description": "Seattle's Best Coffee, Post Alley Blend, 1lb",
-      "image": "https://m.media-amazon.com/images/I/71speMYV5UL._AC_UL320_.jpg",
-      "price": 8.49,
-      "in_stock_qty": 10
-    },
-    {
       "id": "29",
-      "product_id": "11",
       "product_type": "coffee",
       "brand_name": "Bones Coffee",
       "description": "Bones Coffee Company, Maple Bacon, Whole Bean, 1lb",
       "image": "https://m.media-amazon.com/images/I/71gM2UWsIoL._AC_UL320_.jpg",
       "price": 17.99,
-      "in_stock_qty": 6
+      "in_stock_qty": 12
     },
     {
       "id": "8ed3",
-      "product_id": "12",
       "product_type": "coffee",
       "brand_name": "Kitty Town Coffee",
       "description": "Vanilla Buttercream, Whole Bean, 2lb",
@@ -81,40 +19,18 @@
       "in_stock_qty": 3
     },
     {
-      "id": "0d2f",
-      "product_id": "13",
+      "id": "12d4",
+      "brand_name": "Bones Coffee Company",
       "product_type": "coffee",
-      "brand_name": "Chamberlain Coffee",
-      "description": "Fluffy Lamb, Vanilla Cold Brew Blend 1lb",
-      "image": "https://encrypted-tbn1.gstatic.com/shopping?q=tbn:ANd9GcTofqyra9ghC7MIYei3uNmKIJmCmHm9x9BRNVPh6bQ4BkiyvMqSWqcU_A4ZPNe0HFIt-6m5H2AT6FN-XUvespXpXKKisvIAlGQsXWgX43GvQtT3GOcnTfAh&usqp=CAE",
-      "price": 12,
-      "in_stock_qty": 5
-    },
-    {
-      "id": "4cbb",
-      "product_id": "4562",
-      "product_type": "coffee",
-      "brand_name": "Stone Street",
-      "description": "Blueberry Cobbler, Medium Roast, 1lb",
-      "image": "https://encrypted-tbn1.gstatic.com/shopping?q=tbn:ANd9GcRMMVZY4EkTuQrx68ihP-9VchzAk03iMKbW-rJzmOGmtWk3kq1xPu0nVGXrGVdnQSLfSDA8uribEBPw-yeJGqxGrh7IEhifAzj8IL21RlM&usqp=CAE",
-      "price": 17.99,
-      "in_stock_qty": 4
-    },
-    {
-      "id": "acbb",
-      "product_id": "734",
-      "product_type": "coffee",
-      "brand_name": "Verve",
-      "description": "Medium Roast, Specialty Coffee, 1lb",
-      "image": "https://encrypted-tbn0.gstatic.com/shopping?q=tbn:ANd9GcT-udGZBw38pEzKy1mU9Hg4ZXmZhnU9tMYsZIqBpmlA_xmQ0zoKzq168Ifz9tvNtnhabx9ykjJR1vRpvuNdTl1geEyIaQGoBhSBdCyFLo6ESMpwVd45eXJ8&usqp=CAE",
-      "price": 10,
-      "in_stock_qty": 1
+      "description": "Cookies 'N Dreams Ground Gourmet Coffee 12oz",
+      "image": "https://encrypted-tbn0.gstatic.com/shopping?q=tbn:ANd9GcR_D7cEbKPxSllgNUM8GY-A65sEQgo04zMiGGw9NQncCvDgZFMytYmfoUN7aCGBdLVU89AQscI1mzEuUSfIYQ5kFb9irAtBFm8C39s08NqCiGaCGGZsgnc99w&usqp=CAE",
+      "price": 15.99,
+      "in_stock_qty": 7
     }
   ],
   "equipment": [
     {
       "id": "2",
-      "product_id": "14",
       "product_type": "equipment",
       "brand_name": "Breville",
       "description": "Breville Barista Express Espresso Machine",
@@ -124,55 +40,22 @@
     },
     {
       "id": "6",
-      "product_id": "16",
       "product_type": "equipment",
       "brand_name": "Nespresso",
       "description": "Nespresso VertuoPlus Coffee and Espresso Maker",
       "image": "https://m.media-amazon.com/images/I/71zI-ry-gXL._AC_UL320_.jpg",
       "price": 149.99,
-      "in_stock_qty": 7
+      "in_stock_qty": 11
     },
     {
       "id": "8",
-      "product_id": "17",
       "product_type": "equipment",
       "brand_name": "Chemex",
       "description": "Chemex Pour-Over Glass Coffeemaker",
       "image": "https://m.media-amazon.com/images/I/51rEZwQAmqL._AC_UL320_.jpg",
       "price": 44.99,
-      "in_stock_qty": 5
+      "in_stock_qty": 13
     }
   ],
-  "cart": [
-    {
-      "id": "5a5b",
-      "product_id": "16",
-      "product_type": "equipment",
-      "brand_name": "Nespresso",
-      "description": "Nespresso VertuoPlus Coffee and Espresso Maker",
-      "image": "https://m.media-amazon.com/images/I/71zI-ry-gXL._AC_UL320_.jpg",
-      "price": 149.99,
-      "in_stock_qty": 7
-    },
-    {
-      "id": "57f3",
-      "product_id": "17",
-      "product_type": "equipment",
-      "brand_name": "Chemex",
-      "description": "Chemex Pour-Over Glass Coffeemaker",
-      "image": "https://m.media-amazon.com/images/I/51rEZwQAmqL._AC_UL320_.jpg",
-      "price": 44.99,
-      "in_stock_qty": 5
-    },
-    {
-      "id": "b0e5",
-      "product_id": "1",
-      "product_type": "coffee",
-      "brand_name": "Lavazza",
-      "description": "Lavazza Super Crema Whole Bean Coffee Blend, 2.2lb",
-      "image": "https://m.media-amazon.com/images/I/51MMPznBM1L._AC_UL320_.jpg",
-      "price": 23.99,
-      "in_stock_qty": 10
-    }
-  ]
+  "cart": []
 }
