+++ my-app/src/components/CoffeeCard.js
@@ -1,22 +1,11 @@
 import React from "react"
-import Button from "react-bootstrap/Button"
 import Card from "react-bootstrap/Card"
 import DeleteCoffee from "./DeleteCoffee"
+import AddCoffeeToCart from "./AddCoffeeToCart"
 import "../App.css"
 
 function CoffeeCard({ coffeeListing, onAddToCart, onDeleteCoffee }) {
   const { description, image, price, in_stock_qty } = coffeeListing
-  const { id, ...newCartItem } = coffeeListing
-
-  const handleAddToCart = () => {
-    fetch("http://localhost:3001/cart", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(newCartItem),
-    }).then(onAddToCart)
-  }
 
   return (
     <Card className="h-100">
@@ -36,9 +25,10 @@ function CoffeeCard({ coffeeListing, onAddToCart, onDeleteCoffee }) {
         <Card.Text>
           Price: ${price} | Qty: {in_stock_qty}
         </Card.Text>
-        <Button size="sm" variant="primary" onClick={handleAddToCart}>
-          Add To Cart
-        </Button>
+        <AddCoffeeToCart
+          coffeeListing={coffeeListing}
+          onAddToCart={onAddToCart}
+        />
       </Card.Body>
     </Card>
   )
